ChatGPT v1
We are going to make an HTML file for designing 1-bit (i.e., each
"pixel" can only be black or white) images of any size. We'll eventually
have a grid the user interacts with, the ability to provide a reference
image and the ability to export to image to a blown up version. Let's
start with the input grid. It needs to be of a configurable size. When
you click on one of the blocks, it will switch from dark grey to white
or vice-versa depending on how it is currently set.