ChatGPT v5
Perfect. Now let's move to another part of the problem. We need to allow
the user to provide a reference image. When provided, it should be the
background of the cell grid and the different cells should have an
opacity of 70%. We'll also need to tightly pack our grid of squares.