ChatGPT v8 (and me)
I have the following HTML that is settings for a wider page. Let's me it
look pretty like putting it in a box and ensuring the labels for each
input are to the left of their inputs.

```
INSERT CODE HERE
```

---

I also made some modifications myself to make these styling changes
easier.