ChatGPT v9 (and me)
I have a sequence of divs in columns and rows, but they are not
compactly presented when there aren't enough of them to fill the width
of the page. I've found adjusting the width of their parent div to 0%
makes them tightly packed, but any background image I provide is the not
visible. What's a more precise way to configure the width to only be
what is necessary for the contents? Here is the HTML for the divs:

```
INSERT CODE HERE
```

---

Then I played around with it a bit myself to get to the current setting.